Forked from SamplePlugin and renamed to Dancingway. Added appropriate renames and description.
