+++ Dancingway/Dancingway.csproj
@@ -4,9 +4,9 @@
     <Authors></Authors>
     <Company></Company>
     <Version>0.0.0.1</Version>
-    <Description>A sample plugin.</Description>
+    <Description>A plugin to provide a dancing roulette.</Description>
     <Copyright></Copyright>
-    <PackageProjectUrl>https://github.com/goatcorp/SamplePlugin</PackageProjectUrl>
+    <PackageProjectUrl>https://github.com/calli-yama/Dancingway</PackageProjectUrl>
   </PropertyGroup>
 
   <PropertyGroup>
