+++ Dancingway/Windows/ConfigWindow.cs
@@ -1,9 +1,9 @@
-﻿using System;
+using System;
 using System.Numerics;
 using Dalamud.Interface.Windowing;
 using ImGuiNET;
 
-namespace SamplePlugin.Windows;
+namespace Dancingway.Windows;
 
 public class ConfigWindow : Window, IDisposable
 {
