+++ Dancingway/Windows/MainWindow.cs
@@ -1,10 +1,10 @@
-﻿using System;
+using System;
 using System.Numerics;
 using Dalamud.Interface.Windowing;
 using ImGuiNET;
 using ImGuiScene;
 
-namespace SamplePlugin.Windows;
+namespace Dancingway.Windows;
 
 public class MainWindow : Window, IDisposable
 {
